updated scores page

+++ Trivia Project/Backend/include/managers/GameManager/GameData.h
@@ -5,6 +5,6 @@ struct GameData
 {
 	unsigned int correctAnswerCount;
 	unsigned int wrongAnswerCount;
-	unsigned int AverageAnswerTime;
+	float AverageAnswerTime;
 };
